added rsvp modal to event details and dashboard, work in progress

+++ event-finder-API/src/main/java/org/launchcode/event_finder/Repositories/RSVPRepository.java
@@ -8,7 +8,7 @@ import java.util.Optional;
 
 public interface RSVPRepository extends JpaRepository<RSVP, Long> {
     List<RSVP> findByEventId(Long eventId);
-    List<RSVP> findByUserId(Long userId);
+    List<RSVP> findByUserId(Integer userId);
     Optional<RSVP> findByUserIdAndEventId(Integer userId, Long eventId);
 }
